worker/pyproject.toml
[project]
name = "learn-plattform-worker"
version = "0.1.0"
description = "Scheduled jobs for learn-plattform"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.10.0",
    "supabase>=2.10.0",
    "openai>=1.55.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.8.0",
    "black>=24.10.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
target-version = ["py312"]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["jobs", "scrapers"]
